backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
supabase>=2.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest-mock>=3.14.0
